Change threshold instead of switch for denoise
Previously switch was used to turn denoise off and on, this change
means that the button will instead change the threshold. To do this
the notebook is updated to generated a header file with the two
thresholds. app_dsp.c is updated to send these values to the noise
suppressor.